Code quality fixes.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@8742 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/command.cc
@@ -1013,10 +1013,9 @@ static std::vector<std::string> _get_god_keys()
 {
     std::vector<std::string> names;
 
-    for (int i = ((int) GOD_NO_GOD) + 1; i < NUM_GODS; i++)
+    for (int i = GOD_NO_GOD + 1; i < NUM_GODS; i++)
     {
         god_type which_god = static_cast<god_type>(i);
-
         names.push_back(god_name(which_god));
     }
